Added insert sql script

+++ 3.db-project-advanced/ELearningPortalDB/Tables/courses.sql
@@ -6,6 +6,6 @@
 	[updatedOn] [date] NULL,
 	[isPrivate] [bit] NULL,
 	[totalViewingCounter] [int] NULL,
-	[categories_id] [int] NOT NULL,
+	[categories_id] [tinyint] NOT NULL,
 	CONSTRAINT [FK_courses_categories] FOREIGN KEY ([categories_id]) REFERENCES [dbo].[categories] ([id])
 )
